+++ src/Runtime/Common/quaternionToAxisAngle.ts
@@ -1,8 +1,5 @@
-import type { Vector3 } from "@babylonjs/core/Maths/math.vector";
-import { Quaternion } from "@babylonjs/core/Maths/math.vector";
+import type { Quaternion, Vector3 } from "@babylonjs/core/Maths/math.vector";
 import type { DeepImmutable } from "@babylonjs/core/types";
-
-const tempQuaternion = new Quaternion();
 
 /**
  * Quaternion to axis angle
@@ -12,13 +9,14 @@ const tempQuaternion = new Quaternion();
  * @returns angle
  */
 export function quaternionToAxisAngle(q: DeepImmutable<Quaternion>, outAxis: Vector3): number {
-    const quaternion = tempQuaternion.copyFrom(q).normalize();
-    const angle = 2 * Math.acos(quaternion.w);
-    const s = Math.sqrt(1 - quaternion.w * quaternion.w);
-    if (s < 0.001) {
-        outAxis.set(quaternion.x, quaternion.y, quaternion.z).normalize();
+    outAxis.set(q.x, q.y, q.z);
+    const length = outAxis.length();
+    if (length >= 1.0e-8) {
+        const angle = 2.0 * Math.atan2(length, q.w);
+        outAxis.scaleInPlace(1.0 / length);
+        return angle;
     } else {
-        outAxis.set(quaternion.x / s, quaternion.y / s, quaternion.z / s);
+        outAxis.set(1, 0, 0);
+        return 0;
     }
-    return angle;
 }